orchestrator/requirements.txt
fastapi
uvicorn
pydantic
requests
opencv-python 
pillow
